py2-3 compatibility, file organization

+++ mm3_ChannelPicker.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
-from __future__ import print_function
+from __future__ import print_function, division
+import six
 
 # import modules
 import sys
@@ -24,6 +25,7 @@ plt.rcParams['axes.linewidth']=0.5
 
 from skimage.exposure import rescale_intensity # for displaying in GUI
 from scipy.misc import imresize
+from skimage.external import tifffile as tiff
 import multiprocessing
 from multiprocessing import Pool
 import warnings
@@ -42,11 +44,6 @@ cmd_subfolder = os.path.realpath(os.path.abspath(
                                  inspect.currentframe()))[0], "external_lib")))
 if cmd_subfolder not in sys.path:
     sys.path.insert(0, cmd_subfolder)
-
-# supress the warning this always gives
-with warnings.catch_warnings():
-    warnings.simplefilter("ignore")
-    import tifffile as tiff
 
 # this is the mm3 module with all the useful functions and classes
 import mm3_helpers as mm3
@@ -403,7 +400,7 @@ if __name__ == "__main__":
     channel_masks = mm3.load_channel_masks()
 
     # make list of FOVs to process (keys of channel_mask file), but only if there are channels
-    fov_id_list = sorted([fov_id for fov_id, peaks in channel_masks.items() if peaks])
+    fov_id_list = sorted([fov_id for fov_id, peaks in six.iteritems(channel_masks) if peaks])
 
     # remove fovs if the user specified so
     if (len(user_spec_fovs) > 0):
@@ -456,8 +453,8 @@ if __name__ == "__main__":
             mm3.information("Finished cross correlations for FOV %d." % fov_id)
 
         # get results from the pool and put the results in the dictionary if succesful
-        for fov_id, peaks in crosscorrs.items():
-            for peak_id, result in peaks.items():
+        for fov_id, peaks in six.iteritems(crosscorrs):
+            for peak_id, result in six.iteritems(peaks):
                 if result.successful():
                     # put the results, with the average, and a guess if the channel
                     # is full into the dictionary
@@ -485,9 +482,9 @@ if __name__ == "__main__":
         # if there is cross corrs, use it. Otherwise, just make everything -1
         if crosscorrs:
             # update dictionary on initial guess from cross correlations
-            for fov_id, peaks in crosscorrs.items():
+            for fov_id, peaks in six.iteritems(crosscorrs):
                 specs[fov_id] = {}
-                for peak_id, xcorrs in peaks.items():
+                for peak_id, xcorrs in six.iteritems(peaks):
                     # update the guess incase the parameters file was changed
                     xcorrs['full'] = xcorrs['cc_avg'] < p['channel_picker']['channel_picking_threshold']
 
@@ -496,7 +493,7 @@ if __name__ == "__main__":
                     else: # default to don't analyze
                         specs[fov_id][peak_id] = -1
         else: # just set everything to 1 and go forward.
-            for fov_id, peaks in channel_masks.items():
+            for fov_id, peaks in six.iteritems(channel_masks):
                 specs[fov_id] = {peak_id: 1 for peak_id in peaks.keys()}
     else:
         mm3.information('Loading supplied specifiication file.')
@@ -522,7 +519,12 @@ if __name__ == "__main__":
                                       outputdir=outputdir, phase_plane=p['phase_plane'])
 
     # Save out specs file in yaml format
-    with open(os.path.join(ana_dir,"specs.yaml"), 'w') as specs_file:
-        yaml.dump(data=specs, stream=specs_file, default_flow_style=False, tags=None)
+    if not os.path.isfile(os.path.join(ana_dir, 'specs.yaml')):
+        with open(os.path.join(ana_dir, 'specs.yaml'), 'w') as specs_file:
+            yaml.dump(data=specs, stream=specs_file, default_flow_style=False, tags=None)
+    else:
+        mm3.warning('specs.yaml file already exists in analysis folder. Saving to specs_1.pkl')
+        with open(os.path.join(ana_dir, 'specs_1.yaml'), 'w') as specs_file:
+            yaml.dump(data=specs, stream=specs_file, default_flow_style=False, tags=None)
 
-    mm3.information("Finished.")
+    mm3.information('Finished.')
